:recycle: Refactor: hooks 파일 이동
- hooks 폴더 구조 세분화 (util, common, api)
