+++ src/application/hooks/common/usePopup.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from 'react';
 import { useCallback } from 'react';
 
-import useToast from '@/application/hooks/useToast';
+import useToast from '@/application/hooks/common/useToast';
 import Popup from '@/components/common/Popup';
 
 const usePopup = () => {
