+++ src/application/hooks/utils/useClickOutside.ts
